requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types 
antlr4-python3-runtime==4.9.3
anyio
argon2-cffi
argon2-cffi-bindings
astropy==5.2.2
async-timeout==5.0.1
cachetools==4.2.4
coloredlogs==15.0.1
comm
docker-pycreds==0.4.0
dominate==2.6.0
easydict==1.13
eval_type_backport==0.2.2
face-alignment==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
graphviz==0.8.4
grpcio==1.68.1
humanfriendly==10.0
imageio==2.35.1
iopath==0.1.10
ipython
jedi
Jinja2==3.1.5
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
mdurl
mistune
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
mxnet==1.9.1
networkx==3.1
ninja==1.11.1.3
numpy==1.23.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
olefile
omegaconf==2.3.0
onnx==1.17.0
onnx2torch==1.5.15
onnxruntime==1.19.2
onnxsim==0.4.36
opencv-python==3.4.9.33
pandas
pillow==10.4.0
prettytable==3.11.0
propcache==0.2.0
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyDeprecate==0.3.1
pyerfa==2.0.0.3
pytz
PyWavelets==1.4.1
PyYAML==6.0.2
requests
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.16.2
scikit-learn
scipy==1.10.1
sentry-sdk==2.19.2
setproctitle==1.3.4
smmap==5.0.2
sphinx_glpi_theme==0.6
sympy==1.13.3
tabulate==0.9.0
tensorboard==1.15.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.4.0
tifffile==2023.7.10
torchmetrics==0.6.0
tqdm
trimesh==3.9.20
triton==3.0.0
wandb==0.19.1
Werkzeug==3.0.6
yacs==0.1.8
yarl==1.15.2
